+++ src/main/java/cache/LocalCacheSupport.java
@@ -1,6 +1,5 @@
 package cache;
 
-import static cache.CacheStorage.*;
 import static cache.type.LocalCacheType.*;
 import static org.apache.commons.lang3.ObjectUtils.*;
 
@@ -11,18 +10,40 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-import org.apache.commons.lang3.ObjectUtils;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.annotation.Pointcut;
+import org.aspectj.lang.reflect.MethodSignature;
 import cache.impl.EnumMapCacheStorage;
 import cache.impl.MapCacheStorage;
-import cache.type.CommonCacheType;
 import cache.type.LocalCacheType;
 import util.StringUtil;
 
+@Aspect
 public class LocalCacheSupport {
 	private final ThreadLocal<EnumMap<LocalCacheType, CacheStorage>> threadLocalCache = new InheritableThreadLocal<>();
 
-	public <V> V getCache(CommonCacheType key) {
-		return getCache(COMMON_CACHE, key);
+	@Pointcut("@annotation(target) && execution(* *(..))")
+	public void annotion(LocalCacheable target) {
+	}
+
+	@SuppressWarnings("unchecked")
+	@Around("annotion(target)")
+	public <T> T methodCall(ProceedingJoinPoint invoker, LocalCacheable target) throws Throwable {
+		String key = generateCacheKey(target.keyFormat(), target.keyPrefix(), invoker.getArgs(), ((MethodSignature) invoker.getSignature()).getMethod().getParameterAnnotations());
+
+		T cachedValue = getCache(target.type(), key);
+		if (cachedValue == null) {
+			cachedValue = (T) invoker.proceed();
+			setCache(target.type(), target.maxSize(), key, cachedValue);
+		}
+
+		return cachedValue;
+	}
+
+	public <V> V getCache(LocalCacheType key) {
+		return getCache(SINGLETON_CACHE, key);
 	}
 
 	@SuppressWarnings("unchecked")
@@ -36,8 +57,8 @@ public class LocalCacheSupport {
 		return cacheStorage.getCache(key);
 	}
 
-	public <V> void setCache(CommonCacheType key, V val) {
-		setCache(COMMON_CACHE, INF, key, val);
+	public <V> void setCache(LocalCacheType key, V val) {
+		setCache(SINGLETON_CACHE, null, key, val);
 	}
 
 	@SuppressWarnings("unchecked")
@@ -48,13 +69,17 @@ public class LocalCacheSupport {
 
 		if (cacheStorage == null) {
 			int cacheSize = defaultIfNull(size, type.getDefaultSize());
-			cacheStorage = type == COMMON_CACHE ? new EnumMapCacheStorage<>(key, cacheSize) : new MapCacheStorage<>(cacheSize);
+			cacheStorage = type == SINGLETON_CACHE ? new EnumMapCacheStorage<>(key, cacheSize) : new MapCacheStorage<>(cacheSize);
 			cacheStorageCollection.put(type, cacheStorage);
 		}
 
 		cacheStorage.setCache(key, val);
 	}
 
+	/**
+	 *
+	 * @return
+	 */
 	private EnumMap<LocalCacheType, CacheStorage> getCacheStorageCollection() {
 		EnumMap<LocalCacheType, CacheStorage> cacheStorageCollection = threadLocalCache.get();
 		if (cacheStorageCollection == null) {
@@ -73,7 +98,7 @@ public class LocalCacheSupport {
 	 * @param annotations 메서드 파라미터에 적용되어있는 어노테이션 리스트
 	 * @return 캐시키
 	 */
-	public String generateCacheKey(String keyForamt, String keyPrefix, Object[] args, Annotation[][] annotations) {
+	private String generateCacheKey(String keyForamt, String keyPrefix, Object[] args, Annotation[][] annotations) {
 		//메서드 파라미터중, @CacheKey 어노테이션이 적용되어있는 파라미터만 키에 포함시킨다.
 		List<Object> keyParamList = IntStream.range(0, args.length)
 				.boxed()
