--- src/main/java/interceptor/ThreadLocalCacheInterceptor.java
@@ -1,36 +1,0 @@
-package interceptor;
-
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.annotation.Around;
-import org.aspectj.lang.annotation.Aspect;
-import org.aspectj.lang.annotation.Pointcut;
-import org.aspectj.lang.reflect.MethodSignature;
-import cache.LocalCacheSupport;
-import cache.LocalCacheable;
-
-@Aspect
-public class ThreadLocalCacheInterceptor {
-	private LocalCacheSupport localCacheSupport;
-
-	public void setLocalCacheSupport(LocalCacheSupport localCacheSupport) {
-		this.localCacheSupport = localCacheSupport;
-	}
-
-	@Pointcut("@annotation(target) && execution(* *(..))")
-	public void annotion(LocalCacheable target) {
-	}
-
-	@SuppressWarnings("unchecked")
-	@Around("annotion(target)")
-	public <T> T methodCall(ProceedingJoinPoint invoker, LocalCacheable target) throws Throwable {
-		String key = localCacheSupport.generateCacheKey(target.keyFormat(), target.keyPrefix(), invoker.getArgs(), ((MethodSignature) invoker.getSignature()).getMethod().getParameterAnnotations());
-
-		T cachedValue = localCacheSupport.getCache(target.type(), key);
-		if (cachedValue == null) {
-			cachedValue = (T) invoker.proceed();
-			localCacheSupport.setCache(target.type(), target.maxSize(), key, cachedValue);
-		}
-
-		return cachedValue;
-	}
-}
